+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Infrastructure maintenance:
   * Create issue template for bug reports.
   * Create issue template for feature requests.
+  * Created internal settings storage with SharedPreferencesWithCache package.
 
 ## 0.2.34 - 2026-02-05 - The [Viola Caipira](https://pt.wikipedia.org/wiki/Viola_caipira) release
 * New features: